refactor(xlites): moved share DSC / audio jack code into own files
