+++ radio/src/targets/simu/simpgmspace.cpp
@@ -548,23 +548,7 @@ void calcConsumption()
 {
 }
 
-#if defined(HEADPHONE_TRAINER_SWITCH_GPIO)
-void enableHeadphone()
-{
-}
-
-void enableTrainer()
-{
-}
-
-void enableSpeaker()
-{
-}
-
-void disableSpeaker()
-{
-}
-#endif
+void handleJackConnection() {}
 
 int trainerModuleSbusGetByte(unsigned char*) { return 0; }
